new style laundry and resturant

+++ app/Http/Controllers/RoomController.php
@@ -7,9 +7,11 @@ use App\Models\BillDeta;
 use App\Models\Clothe;
 use App\Models\Guest;
 use App\Models\Institution;
+use App\Models\Ledger;
 use App\Models\Meal;
 use App\Models\Room;
 use App\Models\RoomPrice;
+use App\Models\SubAccount;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -23,6 +25,7 @@ class RoomController extends Controller
         $clothes = Clothe::latest()->get();
         $roomprices = RoomPrice::latest()->get();
         $institutions = Institution::latest()->get();
+        $accounts = SubAccount::latest()->get();
         $rooms = Room::where([
             ['number', '!=', Null],
             [function ($query) use ($request) {
@@ -34,7 +37,7 @@ class RoomController extends Controller
             ->with(['guest', 'roomprice', 'partner', 'institution', 'user', 'bill'])
             ->orderBy("id", "asc")
             ->paginate(8);
-        return view('rooms.index', compact(['rooms', 'guests', 'roomprices', 'institutions', 'meals', 'clothes', 'roomall']))
+        return view('rooms.index', compact(['rooms', 'accounts', 'guests', 'roomprices', 'institutions', 'meals', 'clothes', 'roomall']))
             ->with(`i`, (request()->input('page', 1) - 1) * 5);
     }
 
@@ -53,7 +56,7 @@ class RoomController extends Controller
         $room->user_id = Auth::user()->id;
         $room->save();
 
-        return redirect('/rooms')->with('success', 'تم إضافة غرفة جديدة');
+        return redirect()->back()->with('success', 'تم إضافة غرفة جديدة');
     }
 
     public function update(Request $request, room $room)
@@ -75,14 +78,18 @@ class RoomController extends Controller
         // check if the partner id is = to guest id
         if ($request->input('partner_id')) {
             if ($request->input('partner_id') == $request->input('guest_id')) {
-                return redirect('/rooms')->with('warning', 'لا يمكن اختيار النزيل كمرافق');
+                return redirect()->back()->with('warning', 'لا يمكن اختيار النزيل كمرافق');
             }
         }
 
 
         // Rent Room
+
         // check if the opration is rent room or not
         if ($request->input('guest_id')) {
+            if ($request->input('guest_id') == null or $request->input('roomprice_id') == null or $request->input('institution_id') == null) {
+                return redirect()->back()->with('error1', 'هناك بعض الحقول الخالية ');
+            }
             // Make sure that guest number is exist in guest tabel
             if (
                 Guest::where('id', '=', $request->input('guest_id'))->exists() and
@@ -93,39 +100,42 @@ class RoomController extends Controller
 
                 // dd(Room::where('guest_id', '=', $request->input('guest_id'))->exists());
                 // Update Room
-                $room->number = $request->input('number');
-                $room->status = $request->input('status');
+
                 if (!Room::where('guest_id', '=', $request->input('guest_id'))->exists()) {
                     if (!Room::where('partner_id', '=', $request->input('guest_id'))->exists()) {
                         $room->guest_id = $request->input('guest_id');
                     } else {
-                        return redirect('/rooms')->with('error1', 'هذا النزيل مرافق في غرفة اخرى ');
+                        return redirect()->back()->with('error1', 'هذا النزيل مرافق في غرفة اخرى ');
                     }
                 } else {
-                    return redirect('/rooms')->with('error1', 'هذا النزيل ساكن في غرفة اخرى ');
+                    return redirect()->back()->with('error1', 'هذا النزيل ساكن في غرفة اخرى ');
                 }
-                $room->roomprice_id = $request->input('roomprice_id');
+
                 if ($request->input('partner_id')) {
                     if (Guest::where('id', '=', $request->input('partner_id'))->exists()) {
                         if (!Room::where('partner_id', '=', $request->input('partner_id'))->exists()) {
                             if (!Room::where('guest_id', '=', $request->input('partner_id'))->exists()) {
                                 $room->partner_id = $request->input('partner_id');
                             } else {
-                                return redirect('/rooms')->with('error1', 'هذا المرافق ساكن في غرفة اخرى ');
+                                return redirect()->back()->with('error1', 'هذا المرافق ساكن في غرفة اخرى ');
                             }
                         } else {
-                            return redirect('/rooms')->with('error1', 'هذا النزيل مرافق في غرفة اخرى ');
+                            return redirect()->back()->with('error1', 'هذا النزيل مرافق في غرفة اخرى ');
                         }
                     } else {
-                        return redirect('/rooms')->with('error1', 'رقم المرافق غير موجود ');
+                        return redirect()->back()->with('error1', 'رقم المرافق غير موجود ');
                     }
                 }
                 if (Institution::where('id', '=', $request->input('institution_id'))->exists()) {
                     $room->institution_id = $request->input('institution_id');
                 } else {
-                    return redirect('/rooms')->with('error1', 'لا توجد مؤسسة بهذا الرقم ');
+                    return redirect()->back()->with('error1', 'لا توجد مؤسسة بهذا الرقم ');
                 }
                 $room->user_id = Auth::user()->id;
+                $room->roomprice_id = $request->input('roomprice_id');
+                $room->number = $request->input('number');
+                $room->status = $request->input('status');
+                $room->leaving = 1;
 
                 // Bill section
                 $bill = new Bill;
@@ -150,18 +160,21 @@ class RoomController extends Controller
                 $detail->bill_id = $bill->id;
                 $detail->save();
 
-                $room->leaving = 1;
                 $room->save();
 
-                return redirect('/rooms')->with('success', 'تم تسكين النزيل ');
+                return redirect()->back()->with('success', 'تم تسكين النزيل ');
             } else {
-                return redirect('/rooms')->with('error1', 'هناك خطاء في البيانات المدخلة ');
+                return redirect()->back()->with('error1', 'هناك خطاء في البيانات المدخلة ');
             }
         }
 
         // Update Room
         $room->number = $request->input('number');
-        $room->status = $request->input('status');
+        if ($request->input('status') == "ساكنة"){
+            return redirect()->back()->with('error1', 'هناك بعض الحقول الخالية ');
+        }else{
+            $room->status = $request->input('status');
+        }
         $room->guest_id = $request->input('guest_id');
         $room->roomprice_id = $request->input('roomprice_id');
         $room->partner_id = $request->input('partner_id');
@@ -174,25 +187,28 @@ class RoomController extends Controller
 
             //Leaving room by soft deleting bill and bill details
             //Check if post exists before soft deleting
-
-            if (!isset($room->bill)) {
-                return redirect('/rooms')->with('error', 'الفاتورة غير موجودة');
+            if ($room->bill->price != 0) {
+                return redirect()->back()->with('error1', 'لا يمكنك مغادرة الغرفة, لم يتم سداد الفاتورة');
+            } else {
+                if (!isset($room->bill)) {
+                    return redirect()->back()->with('error', 'الفاتورة غير موجودة');
+                }
+                $room->leaving = 0;
+                $room->save();
+                $room->bill->details->each->delete();
+                $room->bill->delete();
+                return redirect()->back()->with('warning', 'لقد قمت بمغادرة نزيل'); // To check if room is empty or not
             }
-            $room->leaving = 0;
-            $room->save();
-            $room->bill->details->each->delete();
-            $room->bill->delete();
-            return redirect('/rooms')->with('warning', 'لقد قمت بمغادرة نزيل'); // To check if room is empty or not
         }
         $room->save();
-        return redirect('/rooms')->with('success', 'تم تحديث الغرفة ');
+        return redirect()->back()->with('success', 'تم تحديث الغرفة ');
     }
 
     # Change Room
     public function changeRoom(Request $request, Room $room)
     {
         $this->validate($request, [
-            'id' => '',
+            'id' => 'required',
             'bill_id' => '',
             'status' => '',
             'guest_id' => '',
@@ -202,10 +218,10 @@ class RoomController extends Controller
             'user_id' => '',
             'leaving' => '',
         ]);
-        
+
         $gole = Room::find($request->input('id'));
         $bill = Bill::find($request->input('bill_id'));
-        
+
         if ($gole->status == 'جاهزة') {
             // Change Room
 
@@ -216,7 +232,7 @@ class RoomController extends Controller
             $gole->institution_id = $request->input('institution_id');
             $gole->user_id = Auth::user()->id;
             $gole->leaving = 1;
-            
+
             $gole->save();
 
             $bill->room_id = $request->input('id');
@@ -229,15 +245,15 @@ class RoomController extends Controller
             $room->institution_id = null;
             $room->user_id = Auth::user()->id;
             $room->leaving = 0;
-            
+
             $room->save();
 
 
-            return redirect('/rooms')->with('success', ' تم تغير الغرفة للنزيل');
-        }elseif ($gole->status == 'ساكنة'){
-            return redirect('/rooms')->with('error1', 'الغرفة المراد النقل اليها ساكنة ');
-        }else{
-            return redirect('/rooms')->with('error1', 'الغرفة غير جاهزة ');
+            return redirect()->back()->with('success', ' تم تغير الغرفة للنزيل');
+        } elseif ($gole->status == 'ساكنة') {
+            return redirect()->back()->with('error1', 'الغرفة المراد النقل اليها ساكنة ');
+        } else {
+            return redirect()->back()->with('error1', 'الغرفة غير جاهزة ');
         }
     }
 
@@ -246,22 +262,134 @@ class RoomController extends Controller
         $this->validate($request, [
             'id' => '',
         ]);
-
-        $room->roomprice_id = $request->input('id');
+        if (RoomPrice::where('id', '=', $request->input('id'))->exists()) {
+            $room->roomprice_id = $request->input('id');
+        } else {
+            return redirect()->back()->with('error1', 'هناك خطأ في سعر الغرفة ');
+        }
         $room->save();
 
-        return redirect('/rooms')->with('success', 'تم تغير سعر الغرفة');
+        return redirect()->back()->with('success', 'تم تغير سعر الغرفة');
+    }
 
+    public function payment(Request $request, Room $room)
+    {
+        $this->validate($request, [
+            'user_id' => '',
+            'guest_id' => 'required',
+            'room_id' => 'required',
+            'statment' => 'required',
+            'price' => 'required',
+            'bill_id' => 'required',
+            'tax' => '',
+            'tourism' => '',
+            'type' => '',
+        ]);
+
+        // Create Detail
+        $detail = new BillDeta;
+        $detail->user_id = Auth::user()->id;
+        $detail->guest_id = $request->input('guest_id');
+        $detail->room_id = $request->input('room_id');
+        $room = Room::find($request->input('room_id'));
+        $detail->statment = $request->input('statment');
+        $detail->bill_id = $request->input('bill_id');
+        $detail->price = $request->input('price');
+        $detail->tax = 0;
+        $detail->tourism = 0;
+        $detail->type = "pay";
+
+        $bill = Bill::find($request->input('bill_id'));
+
+        $this->validate($request, [
+            'statement' => '',
+            'credit' => '', # دائن
+            'debit' => 'required', # مدين
+            'c_amount' => '',
+            'd_amount' => '',
+            'user_id' => '',
+
+        ]);
+
+        // Create Pay
+        $pay = new Ledger();
+        $pay->statement = $request->input('statment');
+        if ($request->input('debit') == 26) {
+            $pay->credit = $request->input('debit');
+            $pay->debit = 21; // حساب الصندوق
+        } 
+        elseif($request->input('debit') == 32) {
+            $pay->credit = $request->input('debit');
+            $pay->debit = 21; // حساب الصندوق
+        }
+        elseif($request->input('debit') == 31) {
+            $pay->credit = $request->input('debit');
+            $pay->debit = 21; // حساب الصندوق
+        }
+        else {
+            $pay->debit = $request->input('debit');
+            $pay->credit = 26; // حساب إيرادات الغرف
+        }
+        $pay->c_amount = $request->input('price');
+        $pay->d_amount = $request->input('price');
+        $pay->user_id = Auth::user()->id;
+
+        // if($room->institution->id != 1 and $request->input('debit') == 26)
+        // {
+        //     return redirect()->back()->with('error1', 'لا يمكنك الدفع كاش في هذه الغرفة');
+        // }
+        // elseif($room->institution->id == 1 and $request->input('debit') != 26)
+        // {
+        //     return redirect()->back()->with('error1', 'لا يمكنك الدفع الدفع على الحساب في هذه الغرفة');
+        // }
+        // if ($bill->price < $request->input('price')) {
+        //     return redirect()->back()->with('error1', 'المبلغ المدخل اكبر من مطالبة الفاتورة');
+        // } else {
+            $bill->price = $bill->price - $detail->price;
+            $detail->save();
+            $bill->save();
+            $pay->save();
+        // }
+
+        return redirect()->back()->with('success', 'تم ' . $detail->statment);
+    }
+
+    public function addPartner(Request $request, Room $room)
+    {
+        $this->validate($request, [
+            'partner_id' => 'required',
+        ]);
+
+        $bill = Bill::find($room->bill->id);
+        $room->partner_id = $request->input('partner_id');
+        $room->save();
+        $bill->partner_id = $request->input('partner_id');
+        $bill->save();
+        return redirect()->back()->with('success', 'تم إضافة مرافق');
+    }
+
+    public function removePartner(Request $request, Room $room)
+    {
+        $this->validate($request, [
+            'partner_id' => '',
+        ]);
+
+        $room->partner_id = null;
+        $room->save();
+        $bill = Bill::find($room->bill->id);
+        $bill->partner_id = null;
+        $bill->save();
+        return redirect()->back()->with('success', 'تمت مغادرة المرافق');
     }
 
     public function destroy(room $room)
     {
         //Check if post exists before deleting
         if (!isset($room)) {
-            return redirect('/rooms')->with('error', 'الغرفة غير موجودة');
+            return redirect()->back()->with('error', 'الغرفة غير موجودة');
         }
 
         $room->delete();
-        return redirect('/rooms')->with('success', 'تم حذف الغرفة');
+        return redirect()->back()->with('success', 'تم حذف الغرفة');
     }
 }
